requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Metrics and Evaluation
torchmetrics>=0.11.0
kornia>=0.7.0

# Data Processing and I/O
numpy>=1.21.0
Pillow>=9.0.0
pandas>=1.3.0

# Configuration and CLI
PyYAML>=6.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
tqdm>=4.62.0
scikit-image>=0.19.0

# Optional: for better numerical stability
scipy>=1.7.0
